refactor: restructure project with module bundling and build configuration for npm package
